erp_neuro_marketing/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
requests==2.32.3
mne==1.6.1
mne-bids==0.15
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
openai==1.40.0
